Split Effects.cs into FilteringEffects, OpenLogEffects, LogReloadEffects, and DatabaseCoordinationEffects
