+++ src/EventLogExpert.Runtime/EventLog/PendingSelectionRestore.cs
@@ -1,0 +1,6 @@
+// // Copyright (c) Microsoft Corporation.
+// // Licensed under the MIT License.
+
+namespace EventLogExpert.Runtime.EventLog;
+
+internal sealed record PendingSelectionRestore(IReadOnlySet<long> SelectedIds, long? SelectedId);
